#17 Add TAKE_SEAT_HANDLER action reducer
